requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
sqlalchemy>=1.4.0
asyncpg>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
aiosmtplib>=1.1.6
jinja2>=3.0.1
fastapi-limiter>=0.1.5
redis>=4.5.1
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
aiosqlite>=0.17.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
alembic>=1.7.0
pytest-cov>=3.0.0

# AI dependencies
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2
torch>=1.9.0
transformers>=4.11.0
tensorflow>=2.8.0,<2.19.0
tf-keras>=2.12.0

# Monitoring and simulation tools
ipaddress>=1.0.23
requests>=2.28.0
watchdog>=2.1.9
argparse>=1.4.0
